Update ThemeDescriptor.swift
Changed LightType to private (eg, defined in the project, not the SDK) enum. Added MinimumVersion, MinimumBuild, and AutoStartDuration.

Also added FileNameParts to return the parts of the contents of the FileName property, and SaveAfterEdit to allow the caller to set the theme to save every time an edit is made.