Update PieceDefinition.swift
Renamed file from PieceDefinition2.swift to PieceDefinition.swift. Also renamed the class PieceDefinition2 to PieceDefinition. Added CustomStringconvertible protocol. Added storage for comment and payload node data. Fixed but in IsUserPiece that always set the value to false. Added ToString to serialized the contents of the instance into an XML node.