Update MainViewController.swift
Added support for the pop-over menu controller. Added code to dump fonts on the system. This is intended to be used only to determine actual font names and should be commented otherwise. Started integrating in-view game buttons. Replaced about box as a storyboard element with the about box as a 3D in-game view object. Added support for animated 3D heartbeat.