Update MainButtonMenuCode.swift
Added flame button which is now repurposed to be a debug handler. When the user pressed the flame button, the pop over gets larger and displays debugging options. When the user pressed the flame button again, the debug options disappear and the pop over shrinks back to its original size.